+++ common/src/main/java/com/faboslav/structurify/world/level/structure/checks/StructureFlatnessCheck.java
@@ -2,12 +2,17 @@ package com.faboslav.structurify.world.level.structure.checks;
 
 import com.faboslav.structurify.common.Structurify;
 import com.faboslav.structurify.common.api.StructurifyStructure;
+import com.faboslav.structurify.common.config.data.StructureData;
 import com.faboslav.structurify.common.config.data.structure.FlatnessCheckData;
+import com.faboslav.structurify.common.util.BiomeUtil;
 import com.faboslav.structurify.world.level.structure.checks.debug.StructureFlatnessCheckOverview;
 import com.faboslav.structurify.world.level.structure.checks.debug.StructureFlatnessCheckSample;
+import net.minecraft.core.HolderSet;
 import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.LevelHeightAccessor;
+import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.chunk.ChunkGenerator;
+import net.minecraft.world.level.levelgen.GenerationStep;
 import net.minecraft.world.level.levelgen.Heightmap;
 import net.minecraft.world.level.levelgen.RandomState;
 import net.minecraft.world.level.levelgen.structure.BoundingBox;
@@ -52,6 +57,25 @@ public final class StructureFlatnessCheck
 		}
 
 		return flatnessCheckDataToCheck;
+	}
+
+	public static boolean canDoFlatnessCheck(
+		StructureCheckData structureCheckData
+	) {
+		var structure = structureCheckData.getStructure();
+		FlatnessCheckData flatnessCheckData = StructureFlatnessCheck.getFlatnessCheckData(structureCheckData);
+
+		if (flatnessCheckData == null || !flatnessCheckData.isEnabled()) {
+			return false;
+		}
+
+		var structureData = structure.structurify$getStructureData();
+
+		if(structureData == null) {
+			return false;
+		}
+
+		return true;
 	}
 
 	public static boolean checkFlatness(